Broke down sync_tests.rs to submodules
